feat: Implement TOML parsing
Adds a simple TOML parser and loads a test file. Renames variables.h to
config.h and moves image files to res folder.

+++ src/vec2.cpp
@@ -2,7 +2,8 @@
 
 vec2::vec2() : x(0.0f), y(0.0f) {}
 
-vec2::vec2(float x, float y) : x(x), y(y) {}
+vec2::vec2(float x, float y)
+    : x(static_cast<float>(x)), y(static_cast<float>(y)) {}
 
 float vec2::dot(const vec2 &other) const { return x * other.x + y * other.y; }
